requirements.txt
gradio>=4.44.0
pinecone-client>=5.0.1
openai>=1.51.0
llama-index>=0.11.0
llama-index-vector-stores-pinecone>=0.3.0
llama-index-embeddings-openai>=0.3.0
llama-index-llms-openai>=0.2.0
tiktoken>=0.7.0
